chore: move all watch logic back into task
